+++ content/scripts/frontend/frontendjs.js
@@ -11,7 +11,7 @@ var eduBookingView = {
     CurrentParticipants: 0,
     DiscountPercent: 0,
     AddParticipant: function () {
-        if (!eduBookingView.SingleParticipant) {
+        if (!eduBookingView.SingleParticipant || eduBookingView.ProgrammeBooking) {
             if (eduBookingView.MaxParticipants === -1 ||
                 eduBookingView.CurrentParticipants <
                     eduBookingView.MaxParticipants) {